+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/channel/list/viewmodel/factory/ChannelListViewModelFactory.kt
@@ -17,8 +17,8 @@ import io.getstream.chat.android.ui.channel.list.viewmodel.ChannelListViewModel
  * @param filter How to filter the channels.
  * @param sort How to sort the channels, defaults to last_updated.
  * @param limit How many channels to return.
- * @param messageLimit The number of messages to fetch for each channel.
- * @param chatEventHandler The instance of [ChatEventHandler] that will be used to handle channel updates event for this combination of [sort] and [filter].
+ * @param memberLimit The number of members per channel.
+ * @param chatEventHandlerFactory The instance of [ChatEventHandlerFactory] that will be used to create [ChatEventHandler].
  *
  * @see Filters
  * @see QuerySort
@@ -28,6 +28,7 @@ public class ChannelListViewModelFactory @JvmOverloads constructor(
     private val sort: QuerySort<Channel> = ChannelListViewModel.DEFAULT_SORT,
     private val limit: Int = 30,
     private val messageLimit: Int = 1,
+    private val memberLimit: Int = 30,
     private val chatEventHandlerFactory: ChatEventHandlerFactory = ChatEventHandlerFactory(),
 ) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
@@ -38,11 +39,12 @@ public class ChannelListViewModelFactory @JvmOverloads constructor(
         @Suppress("UNCHECKED_CAST")
         return ChannelListViewModel(
             ChatDomain.instance(),
-            filter,
-            sort,
-            limit,
-            messageLimit,
-            chatEventHandlerFactory
+            filter = filter,
+            sort = sort,
+            limit = limit,
+            messageLimit = messageLimit,
+            memberLimit = memberLimit,
+            chatEventHandlerFactory = chatEventHandlerFactory,
         ) as T
     }
 }